requirements.txt
certifi==2022.12.7
chardet==4.0.0
Click==8.0.1
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
requests==2.26.0
stripe==3.0.0
toml==0.10.2
Werkzeug==2.0.1
python-dotenv
python-decouple
